@@ -1,6 +1,6 @@
 #if 0
 #elif defined(__arm64__) && __arm64__
-// Generated by Apple Swift version 5.7 (swiftlang-5.7.0.127.4 clang-1400.0.29.50)
+// Generated by Apple Swift version 5.7.1 (swiftlang-5.7.1.135.3 clang-1400.0.29.51)
 #ifndef SWIFTUI_UDF_BINARY_SWIFT_H
 #define SWIFTUI_UDF_BINARY_SWIFT_H
 #pragma clang diagnostic push
